Command Definitions
Commands now contain the default keyword for each command as presented
in the User Manual. Working on I/O to allow users to save custom names.
Help/Await commands will not allow this, as Await should be passively in
the background, and Help should be a fixed and recognizable command.